requirements.txt
aniso8601==3.0.0
APScheduler==3.5.1
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
debtcollector==1.19.0
enum34==1.1.6
Flask==0.12.3
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.0.0
funcsigs==1.0.2
futures==3.2.0
idna==2.6
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
keystoneauth1==3.6.1
ldap3==2.5
MarkupSafe==1.0
monotonic==1.5
msgpack==0.5.6
netaddr==0.7.19
netifaces==0.10.7
os-service-types==1.2.0
oslo.config==6.2.1
oslo.i18n==3.20.0
oslo.serialization==2.25.0
oslo.utils==3.36.2
pbr==4.0.3
psutil==5.4.3
pyasn1==0.4.2
pymongo==3.6.1
PyMySQL==0.8.1
pyparsing==2.2.0
python-keystoneclient==3.16.0
python-swiftclient==3.5.0
pytz==2018.3
PyYAML==3.12
requests==2.21.0
rfc3986==1.1.0
six==1.11.0
SQLAlchemy==1.2.5
stevedore==1.28.0
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
Flask-SocketIO==3.0.1
eventlet==0.23.0
pykafka==2.7.0
python-dateutil==2.7.2
gevent==1.3.3
affinity==0.1.0
numpy==1.15.0
blinker==1.4.0